Code update and test results
Added to code:
-edited output to serial monitor
-Now takes input from monitor for PWM freq
-Made ascii byte to decimal converter

Test results:
The code functioned well. Pin 3 (PWM) was hooked to SV on the BLDC driver, pin 4 (dir) is yet to be used (will need a switching mechanism, ie transistor(s))
255 appears to be the max speed, and 10 is the min noticable speed of the motor.
This is speculatory, as for whatever reason the power supply turns off at certain higher values, and when turned back on the motor spins at an undetermined speed (ie, set to 500, turns off, turn back on motor spins, set to 550 it spins slower)
The power supply also sometimes turns off when set to 100. It mostly happened when moving down from a large number (ie 200), so it may be a large jump in frequency triggers a safety mechanism in the power supply.
Overall the test was successful. Further automation should take the above into account when transitioning speeds.